feat: Migrate SliderSetting and Update UI and Dependencies
This commit migrates the `SliderSetting` composable to the `kmpuiviews` module for multiplatform use, updates its usage across the application, and updates the dependencies.

-   **`SliderSetting` Migration:**
    -   Moved `SliderSetting` from `UIViews` to `kmpuiviews`.
    -   Removed the old `SliderSetting` from `UIViews`.
    - Updated `NotificationSettings`, `GeneralSettings`, `PlaySettings`, `ReaderSheets`, and `OnboardingContent` to use the new `SliderSetting` from `kmpuiviews`.
-   **Dependencies:**
+   - Updated to use `tech.annexflow.compose:constraintlayout-compose-multiplatform` in `kmpuiviews`.
     -   Removed the old `SliderSetting` from the project.
 - **General:**
  - Removed unused imports.
+ - Updated `ReaderSettings` to use the new `SliderSetting`.
+ - Updated imports to be multiplatform.
