+++ kmpuiviews/src/commonMain/kotlin/com/programmersbox/kmpuiviews/presentation/components/SettingsComposables.kt
@@ -28,6 +28,9 @@ import androidx.compose.material3.ProvideTextStyle
 import androidx.compose.material3.RadioButton
 import androidx.compose.material3.RadioButtonColors
 import androidx.compose.material3.RadioButtonDefaults
+import androidx.compose.material3.Slider
+import androidx.compose.material3.SliderColors
+import androidx.compose.material3.SliderDefaults
 import androidx.compose.material3.Switch
 import androidx.compose.material3.SwitchColors
 import androidx.compose.material3.SwitchDefaults
@@ -49,6 +52,8 @@ import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.window.DialogProperties
+import androidx.constraintlayout.compose.ConstraintLayout
+import androidx.constraintlayout.compose.Dimension
 
 @ExperimentalMaterial3Api
 @ExperimentalComposeUiApi
@@ -448,4 +453,87 @@ fun <T> DynamicListSetting(
             ) { dialogPopup.value = true }
             .then(modifier)
     )
+}
+
+@Composable
+fun SliderSetting(
+    sliderValue: Float,
+    updateValue: (Float) -> Unit,
+    range: ClosedFloatingPointRange<Float>,
+    settingTitle: @Composable () -> Unit,
+    modifier: Modifier = Modifier,
+    settingIcon: (@Composable BoxScope.() -> Unit)? = null,
+    settingSummary: (@Composable () -> Unit)? = null,
+    steps: Int = 0,
+    colors: SliderColors = SliderDefaults.colors(),
+    format: (Float) -> String = { it.toInt().toString() },
+    onValueChangedFinished: (() -> Unit)? = null,
+) {
+    ConstraintLayout(
+        modifier = Modifier
+            .padding(8.dp)
+            .fillMaxWidth()
+            .then(modifier)
+    ) {
+        val (
+            icon,
+            info,
+            slider,
+            value,
+        ) = createRefs()
+
+        Box(
+            modifier = Modifier
+                .constrainAs(icon) {
+                    start.linkTo(parent.start)
+                    top.linkTo(parent.top)
+                    bottom.linkTo(parent.bottom)
+                }
+                .padding(8.dp)
+                .padding(end = 16.dp)
+        ) { settingIcon?.invoke(this) }
+
+        Column(
+            modifier = Modifier.constrainAs(info) {
+                top.linkTo(parent.top)
+                end.linkTo(parent.end)
+                start.linkTo(icon.end, margin = 10.dp)
+                width = Dimension.fillToConstraints
+            }
+        ) {
+            ProvideTextStyle(
+                MaterialTheme.typography.bodyLarge.copy(fontWeight = FontWeight.Medium, textAlign = TextAlign.Start)
+            ) { settingTitle() }
+            settingSummary?.let {
+                ProvideTextStyle(MaterialTheme.typography.bodyMedium.copy(textAlign = TextAlign.Start)) { it() }
+            }
+        }
+
+        Slider(
+            value = sliderValue,
+            onValueChange = updateValue,
+            onValueChangeFinished = onValueChangedFinished,
+            valueRange = range,
+            steps = steps,
+            colors = colors,
+            modifier = Modifier.constrainAs(slider) {
+                top.linkTo(info.bottom)
+                end.linkTo(value.start)
+                start.linkTo(icon.end)
+                width = Dimension.fillToConstraints
+            }
+        )
+
+        Text(
+            format(sliderValue),
+            style = MaterialTheme.typography.titleMedium,
+            modifier = Modifier
+                .constrainAs(value) {
+                    end.linkTo(parent.end)
+                    start.linkTo(slider.end)
+                    centerVerticallyTo(slider)
+                }
+                .padding(horizontal = 16.dp)
+        )
+    }
 }
