+++ mangaworld/src/main/java/com/programmersbox/mangaworld/settings/ReaderSettings.kt
@@ -75,13 +75,13 @@ import com.programmersbox.datastore.mangasettings.ReaderType
 import com.programmersbox.kmpuiviews.presentation.components.BackButton
 import com.programmersbox.kmpuiviews.presentation.components.CategorySetting
 import com.programmersbox.kmpuiviews.presentation.components.PreferenceSetting
+import com.programmersbox.kmpuiviews.presentation.components.SliderSetting
 import com.programmersbox.kmpuiviews.presentation.components.SwitchSetting
 import com.programmersbox.mangasettings.MangaNewSettingsHandling
 import com.programmersbox.mangaworld.R
 import com.programmersbox.mangaworld.reader.compose.FloatingBottomBar
 import com.programmersbox.mangaworld.reader.compose.ReaderTopBar
 import com.programmersbox.mangaworld.reader.compose.dpToPx
-import com.programmersbox.uiviews.presentation.components.SliderSetting
 import com.programmersbox.uiviews.utils.ComposableUtils
 import com.programmersbox.uiviews.utils.HideNavBarWhileOnScreen
 import com.programmersbox.uiviews.utils.LightAndDarkPreviews
